requirements.txt
PyYAML~=6.0.2
openpyxl~=3.1.5
openai~=1.82.1
requests~=2.32.3